Implemented service to load images by url
