+++ PasswordManager/Services/FavIconService.cs
@@ -5,7 +5,6 @@ using System.Collections.Concurrent;
 using System.Net.Http;
 using System.Threading;
 using System.Windows.Media;
-using System.Windows.Media.Imaging;
 
 namespace PasswordManager.Services
 {
@@ -15,12 +14,15 @@ namespace PasswordManager.Services
         private readonly ConcurrentDictionary<string, ImageSource> _imagesDict = new();
         private readonly IHttpClientFactory _httpClientFactory;
         private readonly ILogger<FavIconService> _logger;
+        private readonly ImageService _imageService;
 
         public FavIconService(
             IHttpClientFactory httpClientFactory,
-            ILogger<FavIconService> logger)
+            ILogger<FavIconService> logger,
+            ImageService imageService)
         {
             _httpClientFactory = httpClientFactory;
+            _imageService = imageService;
             _logger = logger;
         }
 
@@ -36,16 +38,7 @@ namespace PasswordManager.Services
                 if (_imagesDict.TryGetValue(host, out ImageSource image))
                     return image;
 
-                var client = _httpClientFactory.CreateClient();
-                var request = new HttpRequestMessage(HttpMethod.Get, string.Format(_favIconServiceUrl, host));
-                using var response = client.SendAsync(request, CancellationToken.None).Result;
-                using var stream = response.Content.ReadAsStreamAsync(CancellationToken.None).Result;
-                var bitmapImage = new BitmapImage();
-                bitmapImage.BeginInit();
-                bitmapImage.CacheOption = BitmapCacheOption.OnLoad;
-                bitmapImage.StreamSource = stream;
-                bitmapImage.EndInit();
-                bitmapImage.Freeze();
+                var bitmapImage = _imageService.GetImageAsync(string.Format(_favIconServiceUrl, host), CancellationToken.None).Result;
 
                 _imagesDict.TryAdd(host, bitmapImage);
                 return bitmapImage;
