+++ PasswordManager/Services/ImageService.cs
@@ -1,0 +1,33 @@
+﻿using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+
+namespace PasswordManager.Services
+{
+    public class ImageService
+    {
+        private readonly IHttpClientFactory _httpClientFactory;
+
+        public ImageService(IHttpClientFactory httpClientFactory)
+        {
+            _httpClientFactory = httpClientFactory;
+        }
+
+        public async Task<ImageSource> GetImageAsync(string imageUrl, CancellationToken cancellationToken)
+        {
+            var client = _httpClientFactory.CreateClient();
+            var request = new HttpRequestMessage(HttpMethod.Get, imageUrl);
+            using var response = await client.SendAsync(request, cancellationToken).ConfigureAwait(false);
+            using var stream = await response.Content.ReadAsStreamAsync(cancellationToken).ConfigureAwait(false);
+            var bitmapImage = new BitmapImage();
+            bitmapImage.BeginInit();
+            bitmapImage.CacheOption = BitmapCacheOption.OnLoad;
+            bitmapImage.StreamSource = stream;
+            bitmapImage.EndInit();
+            bitmapImage.Freeze();
+            return bitmapImage;
+        }
+    }
+}
